Track the CP_SET_DRAW_STATE enable mask.
The CP_SET_DRAW_STATE IBs contain bits indicating whether that IB
executes in the BINNING, GMEM and/or SYSMEM pass. These bits need to be
tracked and made available to EmulateStateTracker to know which register
assignments occur in which pass.

This change picks up the bits from CP_SET_DRAW_STATE
(`AdvanceCb()`), passes them to `QueueIB()`, which records them in
`m_ib_queue_enable_mask`. Then in `AdvanceToQueuedIB()`, the value is
loaded into `m_cur_ib_enable_mask` along with the rest of the IB info,
and then passed to `OnIbStart()` via `call_chain_ib_info.m_enable_mask`.
Finally, in `CaptureMetadataCreator::OnIbStart()`, the enable mask is
pushed to `EmulateStateTracker`, which maintains a stack so that the
enable mask of outer IB can be restored.