Flatten PM4 hierarchy for event views
Get rid of all the CP_INDIRECT_BUFFER packets in
event views. Flatten the hierarchy. All packets that
were children of CP_INDIRECT_BUFFER are added to the
event node.

This fixes issue where a CP_INDIRECT_BUFFER that
contains lots of draw calls results in a
CP_INDIRECT_BUFFER packet shown for the 1st draw,
and the rest of the draws contain no PM4 packets.