Fix crash when dealing with IBs where m_skip==true
This is a case where m_start_node_stack assumes that packets
exist in every IB, which is obviously not true for IBs that
are "skipped". No packets get pushed on that stack, and then
it gets popped at the end of OnIbEnd(). This causes a crash
later. Fixed by using the m_new_ib_start flag to determine
if any packets were pushed onto the stack, and not doing
any of the usual things if nothing is pushed.