Fix crash when running capture without a device
Running `dive_client_cli --command capture` without any other parameters
would crash. This is because it wasn't checking whether the rest of the
parameters that it needed were set.

The crash was during trigger_capture. Building target_str for the gRPC
channel requires knowing the device port. Howevever, the device is null
when a serial is not provided; trying to access the port crashes.

The fix is to look at the result of run_package. This returns false when
an error is encountered, like the required params are not provided. It
also seems appropriate to bail if it fails for any other reason since if
the app couldn't be run then profiling results might not be accurate.