Fix crash when running capture without a device (#196)
Running `dive_client_cli --command capture` without any other options crashes. This is because it wasn't checking whether the rest of the options that it needed were set.

The crash was during trigger_capture. Building target_str for the gRPC channel requires knowing the device port. However, the device is null when `--device` is not provided; trying to access the port crashes.

The fix is to look at the result of run_package. This returns false when an error is encountered, like the required options are not provided. We can piggy-back off of the option validation in this case. It also seems appropriate to bail if it fails for any other reason: the app may not be running, so any profiling results colleted might not be accurate.

Additionally:

- Be defensive in trigger_capture and make sure the device is selected.
- Provide a more helpful error output so that the user knows they did something wrong.