Recommend using a virtual environment for Python
On Debian, I found that--while python3 was installed--python was not in
my path. This caused replay to fail.

`python` is hardcoded into DeviceManager::DeployReplayApk as the name of
the Python binary. While we could change this to `python3`, PEP 394
recommends assuming `python` is set appropriately by a virtual
environment to the desired Python version.

As such, recommend that users use a virtual environment for Python.
There are many choices so I leave the particulars of the system
configuration up to the user.