Initialize GFXR capture trigger property
`debug.gfxrecon.capture_android_trigger` is not set on the first launch
of Dive. This causes GFXR to not use triggered trim captures: attempts
to set capture_android_trigger to trigger the capture are ignored.

The GFXR docs (USAGE_android.md) suggest that if you want trimed
capture mode then you need to set the property to something prior to
startup.

Given the observed behavior, always set the property to `false` on
setup. Additionally, clear the property on exit so that other apps
which may want to use GFXR with a full capture don't accidentally
inherit the setting.