Tool to parse .GFXR and craft --dump-resources JSON (#211)
To aid in validation of replay correctness, create a program that processes a GFXR file for all resources that can be used with GFXR `--dump-resources`. This has only been tested on cube_xr, fishtornado_xr, and sample_04_cube; it is likely that additions will need to be made as we test on more applications.