Adding a modification framework to DiveBlockData (#224)
Refactoring DiveBlockData and adding functionality to store and apply
block-level modifications.
- The location of a modification is described with `primary_index` and
`secondary_index`, describing which original block it's anchored to and
how it is positioned in relation to the anchor block (negative for
immediately before, 0 for overwriting, and positive for immediately
afterwards)
- Original block metadata and modification block metadata are stored
separately so that modifications can be manipulated easily without
affecting the original data.
- `BlockVisitor` is defined using the Visitor design pattern, with
`WriterBlockVisitor` made to hold handles to open files (original and
new gfxr files) and write the new file block-by-block in the traversal
order