Run root for OpenXR GFXR capture (#265)
Currently, setting openxr.enable_frame_delimiter requires root. When
running OpenXR GFXR capture, ensure we have root. This should not affect
Vulkan capture on phones.

I moved the setprop to the same location as the root access request so
that the dependency is not spread out in the code base.