Add option to control how many draw calls are dumped
Each draw call can take 2-3 seconds to dump and produces a 31MB file.
Large captures can have many draw calls which increases the dump and
transfer time.

For validation, we only care about the final draw call since it
represents what is shown to the user. This saves a lot of time without
impacting functionality.

To this end, I've added the `--last_draw_only` flag to
gfxr_dump_resources which throws away all except 1 of the draw calls for
each render pass. I've updated replay-with-dump to use that flag.