Leak FDM to avoid crashing from stale image view
There's a Unity bug where a VkImageView is used in a framebuffer which
still references a now-deleted FDM after swapchain recreation. On
vkDestroyImage, GFXR removes all associations between the VkImage and
VkImageView. When GFXR goes to use the attachment in the frame buffer
at the beginning of a render pass, it tries to find the VkImage for the
VkImageView but gets nullptr instead; trying to dereference this causes
a crash.

One workaround for this is to leak FDMs. Since they tend to be fraction
of the extent of the frame buffer, this shouldn't have substantial
negative implications. However, this relies on the FDM being recreated
infrequently.