Fixes for DiveBlockData (#292)
- Fix logic for adding original blocks when the .gfxr file has an
accompanying .gfxa file and the active file changes throughout replay.
  - What seems to happen is that every gfxr "block" that is stored inside
the gfxa file will have a preceeding block in the gfxr file of meta type
`ExecuteBlocksFromFile` that indicates to the file processor to switch
active files to gfxa, and after the block is processed the active file
switches back to the gfxr file.
- Fix the improper handling of `--output_gfxr_path` flag in `host_cli`
that is causing early termination after the file is loaded
- Replace the tester .gfxr file with one that has a separate asset file,
to make our integration test better represent the use case for Dive
- Improve integration test logic (copying setup from
gfxr_dump_resources) to actually check the modified gfxr file's contents