GFXR vkCreateDevice: If family reused, ignore protected queue
When creating a device, you must specify which queues to create. Apps
are allowed to provice a queue family index twice: once for non-
protected queue (flags = 0), and again for protected queue (flags = 1).

For now, we suspect that the protected queue is used for niche
situations. Therefore: replace the assert with a warning, and prefer the
non-protected queue.

Currently, capture doesn't use the flags that are stored so this code
could be removed entirely. Replay uses the flags in some places like the
VulkanResourceInitializer to get a queue on which to submit work.