Revert "Batch resources initialization" (#311)
This reverts commit
https://github.com/LunarG/gfxreconstruct/commit/f296bb1f219ad3d6a2051b8595e48dbe243fd189


https://github.com/LunarG/gfxreconstruct/commit/f296bb1f219ad3d6a2051b8595e48dbe243fd189
introduces a bug. It intends to optimize resource initialization in
playback by reusing a staging buffer and packing many smaller resources
into that buffer; this saves on submits and allocations. However,
certain sequences of meta commands could cause the command buffer and
staging buffer states to fall out of sync and crash the program.

InitImageCommand can result in one of two function calls depending on
the nature of the image:

1. VulkanResourceInitializer::TransitionImage if linear tiling, host-
visible, and with a set layout
2. VulkanResourceInitializer::InitializeImage otherwise

The latter only begins recording if the staging buffer is empty.
However, the former unconditionally stops recording and doesn't touch
the staging buffer. This means that the following will cause a crash:

1. InitImage with a large image to allocate a large staging buffer
2. InitImage with a small image with e.g. optimal tiling. This is staged
in the staging buffer but not submitted.
3. InitImage with a linear tiling, host-visible image. This will submit
the command buffer but not touch the staging buffer.
4. InitImage with a small image with e.g. optimal tiling. Since the
staging buffer is not empty, the command buffer doesn't start recording.
However, it tries to record a command. This crashes.

Revert to avoid this crash