GFXR: Demote --dump-resources dynamic offset check (#323)
Since the spec for vkAllocateDescriptorSets says that "entries that are not used by a pipeline can have undefined
descriptors", replace the assert in vkCmdBindDescriptorSets with a
warning.

If vkUpdateDescriptorSets is not called,
VulkanDescriptorSetBindingInfo.buffer_info will not have contain a
VulkanDescriptorTypeBufferInfo for that binding or array element
(despite the descriptor set layout). This assert should likely take more
context into consideration.