GFXR: If queue family reused, ignore protected queue (#324)
When creating a device, you must specify which queues to create. Apps
are allowed to provide the same a queue family index twice when it wants
both a non-protected queue (flags = 0) and a protected queue (flags =
1).

For now, we suspect that the protected queue is used for niche
situations. Therefore: replace the assert with a warning, and prefer the
non-protected queue.

Currently, capture doesn't use the flags that are stored so this code
could be removed entirely. Replay uses the flags in some places like the
VulkanResourceInitializer to get a queue on which to submit work.