Fixes for GPU time <-> vk correlation (#345)
Fix on Replay apk:
- it seems that the OS is inserting the same empty cmd for left and
right eye without VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT. This in
theory should trigger a validation error, but not in this case
For now we keep both in the frame cache m_frame_cmds, since it is easier
for vk correlation (but the gpu time will be equal to the last cmd that
is being submitted). This does not affect timing since it is empty, but
needs some further investigation.

Fix on vk tree view:
We should not create the tree view by purely following the block order
since
- Some command buffers could be pre-recorded only once, but submitted 2
times in one frame for left and right eye (seens to be allowed in OpenXR
Vulkan apps), in this case, we want to show the vk events in both
submissions
- The app can call vkBeginCommandBuffer on cmdA, then call
vkBeginCommandBuffer on cmdB, then submitt cmdC, then call
vkEndCommandBuffer on cmdB and cmdA. In this case, we should keep the vk
events in submission order (instead of recording order) so that it is
easier to correlate with PM4 and gpu time