Defer releasing fences that are created in setup phase (#358)
We have some resources that are created in the setup phase, but got
released in the frame. When we start to loop the frame, we cannot
release them within the frame. We only push fences in the deferred
release list for now, but might need to add other resources in the
deferred release list in the future.