Ensures user cannot start replay before selecting a device (#368)
Ensures user cannot start replay before selecting a device. The replay
button is disabled until a device is selected. If there is an issue when
setting up the device (disconnection or some other issue), the button is
disabled and the device list is refreshed.

The recording shows the view when a device isn't connected, then
connected, then disconnected during setup.


https://github.com/user-attachments/assets/debf82cf-0589-45e9-84c5-207665105b96