Fixes for CSV artifacts produced by GFXR replay (#374)
Rename `gpu_time.csv` as it is pulled from the device so that the naming
is consistent with other files associated with the same GFXR file, with
the same root and a suffix `_gpu_time.csv`.

Other changes:
- Did the same rename-on-retrieve for the profiling metrics CSV, adding
the suffix `_profiling_metrics.csv`
- Added a check for loading GPU timing info during `LoadFile()` in the
UI
- Refactored `RetrieveTrace()` to `RetrieveFile()` and added some
parameters for clarity
- Some renaming of `path` to `dir` where appropriate for clarity (We
risked bugs such as renaming a file to a specified dir name with no
extension)
- Added UI debug message for GPU timing correlation which will log a
warning if the CSV row count does not match what was found in the GFXR
file.