Allow building Dive UI to log to console (#379)
With WIN32, all logs to stdout are lost. The logs are useful for
debugging.

Add a CMake option() to control whether WIN32_EXECUTABLE is set for the
UI. If WIN32_EXECUTABLE is provided then the app is built as a GUI
(/SUBSYSTEM:WINDOWS) and all logs to stdout are lost.

If you want logs, you can set DIVE_UI_WIN32 to OFF. This will build as
/SUBSYSTEM:CONSOLE and logs to stdout will be printed to a console.