Remove temporary artifacts from previous GFXR replays (#388)
Since we moved to using the GFXR file load folder as the output folder
for replay artifacts, currently there is an issue where subsequent
replays will not have the proper artifacts displayed in the UI because
we have improper cleanup and reload.

This PR addresses the issue for three files (profiling metrics CSV, gpu
timing CSV, and PM4 capture) and adds a warning in the UI for the user
to manually save any temporary files that they don't want cleared on the
next replay. In the future when we have .dive files, more of this will
be hidden from the user for ease of use.

Changes:
- Deleting temp files from earlier replay runs upon initiation of new
replay
- Force reload of GFXR (+rd if present) after replay has completed.
- Adding support for perf counter model reset with no CSV present
- Added replay warning text in UI Analyze Capture menu

Minor changes:
- Refactoring for updating tab views and getting path names
- Extra debug logging around file handling