Use correct STRING type for DIVE_GFXR_GRADLE_CONSOLE
I was noticing CMake warnings:

```
CMake Warning (dev) at CMakeLists.txt:115 (set):
  implicitly converting 'LIST' to 'STRING' type.
This warning is for project developers.  Use -Wno-dev to suppress it.
```

The LIST cache variable type was leftover from a previous implementation. I expect a string (one of gradle `--console` values) so the type should be STRING. This will align the type with the intent as well as remove the warning from the build output.