libwrap: Truncate collected trace file (#404)
The current code tracks separate trace files for each FD. This was
deemed ["useful because they separate traces from OGLES and Vulkan
contexts into different
files"](https://gitlab.freedesktop.org/freedreno/freedreno/-/merge_requests/23#note_1856011)

As a result, collect_trace_file() effectively replaces rd_end() in order
to concatenate all open trace files into a single trace file; this is
easier for the host program to download and analyze.

However, if the target file for concatentation already existed from a
previous trace then it would be appended to. This increases file
transfer times and ruins analysis. To that end, truncate the collected
log file at the start of collect_trace_file().

For simplicity of implementation, this also involved removing the
optimization at the start of append_file.

Note that this is only currently a problem for the PM4 capture UI
feature (`--command=capture`), not the Analyze UI feature
(`--command=profile_replay`). However, this approach should be more
robust to a variety of error conditions compared to some simpler
alternatives.