GFXR: Allocate memory that RenderDoc expects (#430)
RenderDoc replaces external memory with non-external memory when
replaying a .RDC file. If the memory type for the external memory does
not match the non-external memory then RenderDoc throws an error. To
avoid this, if we detect external memory, allocate what we think
RenderDoc wants instead of what's in the .GFXR file. While this allows
RenderDoc to open .RDC files, anything that uses external memory will be
black since RenderDoc doesn't support it.

Based on testing, imported AHardwareBuffers use memoryTypeIndex 1 but
RenderDoc wants it to be 0. We can't call vkGetImageMemoryRequirements
since (apart from dedicated memory) we don't know which image this
memory will be bound to.

Changing memoryTypeIndex in DiveVulkanReplayConsumer is brittle since
the superclass implementation could overwrite memoryTypeIndex. However,
it reduces the number of modifications made to GFXR; this should reduce
the number of merge conflicts when we perform the next subtree pull.

This is protected by a system property until we use the RenderDoc
in-application API to detect if RenderDoc is capturing. This allows the
developer to toggle this behavior if it's causing issues.