Load the RenderDoc API, use in external memory hack
Introduce a function to load the RenderDoc In-Application API.

If the In-Application API can be initialized, this likely means that
the RenderDoc layer has been loaded and is capturing. Use this instead
of the system property for overriding the memoryIndexType for external
memory.

The implementation uses a singleton instance since the way we extend
GFXR right now is not very compatible with dependency injection. I hope
to improve this in the future.

API loading is hidden behind a compile-time swappable implementation in
an attempt to aid unit testing in the future while still maintaining
code readability. The main implementation is in the Android file but a
stub is provided for the convenience of building on other platforms.

The next PR will actually use the API to create a capture.