dive_client_cli: Specify analysis with enum (#442)
Replace separate --dump_pm4 and --enable_perf_counters flags with one
flag that accepts only one value: --gfxr_replay_run_type. Since analysis
options are mutually exclusive, the use of an enum simplifies how the
user specifies the feature and how the CLI validates that the flags are
valid.

As a side-effect, this also allows the CLI to measure GPU time.