Create a RenderDoc capture during replay (#446)
Use the In-Application API to start a RenderDoc capture at the state end
marker and end the capture after all loops during replay. This is hidden
behind a property to avoid interfering with the ability to debug using
the RenderDoc UI. Using a property also avoids the need to modify GFXR
to include a new flag.

The resulting file will be /sdcard/Download/GFXRFILE_capture.rdc. The UI
can use this stable filename scheme to know which file to Download.

It is assumed that if you are launching replay with the Dive flag set
then replay has exclusive access over RenderDoc state; i.e. the user
hasn't also connected with the UI, or launched the app with layer but
without the intent to use Dive-specific behavior.

For simplicity of implementation, the user must already have the
renderdoccmd app installed. If it's not installed, the UI will provide
an error and capture will fail.

Add the RenderDoc capture option to the Analysis window. This will
launch replay with the RenderDoc capture layer enabled to use the API to
create a capture. The RenderDoc capture will be pulled to the same
directory as the GFXR file.