Tell to install RenderDoc with --force-queryable
Through experimentation, it seems tht the use of gpu_debug_layer_app
requires the replay APK to query the renderdoc APK for the layer.
Without --force-queryable, replay will be unable to load the layer and
stop.

The RenderDoc UI will install the APK with this flag. If the user
installs it manually though they just need to make sure they provide
--force-queryable.