fix image capture (#457)
The problem is that std::filesystem::path on Windows creates
Windows-style paths using backslashes (\), but we are sending that path
to an Android shell, which expects forward slashes (/).

When the Android shell sees the backslashes in command: shell screencap
-p /sdcard/Download\gfxr_capture\capture_screenshot.png

It treats the \ as an escape character, effectively "eating" the
separator. The shell command becomes: shell screencap -p
/sdcard/Downloadgfxr_capturecapture_screenshot.png

This creates that single, strangely-named file
(Downloadgfxr_capturecapture_screenshot.png) in the /sdcard/ directory