Allow GFXR replay with validation layer (#479)
Add the --validation_layer CLI option which sets use_validation_layer.
If set, push and enable the Vulkan Validation Layer. This only works for
normal replay to keep things simple; this does not preclude expanding it
to work with other replay modes in the future.

Includes prebuilt Validation Layers for 1.4.328.1. Downloaded from:
https://github.com/KhronosGroup/Vulkan-ValidationLayers/releases/download/vulkan-sdk-1.4.328.1/android-binaries-1.4.328.1.tar.gz

The prebuilt was chosen since compiling from the repo required pulling
in a lot of dependencies. Choosing the prebuilt doesn't preclude us from
going down this path in the future.