Change our cleanup procedure to account for sessions that start with device in a bad state (#508)
Cleanup changes:
- Standardized logging around the start/end of cleanup procedures (and
added a filterable prefix) so that it's easier to follow when reading
the logs
- Added cleanup calls to the constructor for all variants of
applications, so that when the user selects application type it will
trigger a cleanup of app-related things on the device.
- Added cleanup call to the startup of AndroidDevice, so that when the
user selects the device it will trigger a general cleanup of the device
(except for app-related things).
- Renamed `DeviceManager::Cleanup()` to `CleanupPackageProperties()` to
indicate it's app-related cleanup.